scripts/requirements.txt
aiohttp
beautifulsoup4
numpy
pandas
python-dateutil
selenium
bs4
html5lib
requests
webdriver-manager
playwright